Renames to generated code Processor to Handler (#1020)
* Make generated comments look good + some renaming of generated code

* Rename based on recommendations in the issue https://github.com/istio/mixer/issues/1018

* Make the printing of method arguments in the generated code consistent.

Now we just print the data type without the arg name for interface methods.


Former-commit-id: 215cf93bd0e5bc9c4341af120b925a298b4ba785

+++ mixer/tools/codegen/pkg/bootstrapgen/generator_test.go
@@ -17,8 +17,8 @@ package bootstrapgen
 import (
 	"bytes"
 	"io"
-	"io/ioutil"
 	"os"
+	"path"
 	"testing"
 )
 
@@ -48,13 +48,15 @@ func TestGenerator_Generate(t *testing.T) {
 	}
 	for _, v := range tests {
 		t.Run(v.name, func(t *testing.T) {
-			outFile, err := ioutil.TempFile("", v.name)
+			outFile, err := os.Create(path.Join(os.TempDir(), path.Base(v.want)))
 			if err != nil {
 				t.Fatal(err)
 			}
 			defer func() {
-				if removeErr := os.Remove(outFile.Name()); removeErr != nil {
-					t.Logf("Could not remove temporary file %s: %v", outFile.Name(), removeErr)
+				if !t.Failed() {
+					if removeErr := os.Remove(outFile.Name()); removeErr != nil {
+						t.Logf("Could not remove temporary file %s: %v", outFile.Name(), removeErr)
+					}
 				}
 			}()
